ai_service/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
joblib==1.3.2
python-dotenv==1.0.0
